pyproject.toml
[project]
name = "juridiction-senegal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.3.5",
    "fastapi>=0.123.0",
    "fastembed>=0.7.3",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "pydantic>=2.12.5",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "requirements-py>=1.0.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "tiktoken>=0.12.0",
    "transformers>=4.57.3",
    "uvicorn[standard]>=0.38.0",
]
